+++ js/packages/web/src/store/manager.test.ts
@@ -38,11 +38,7 @@ describe("ExecutionManager", () => {
       expect(manager.ctx.status).toBe(TaskStatus.INITIAL);
 
       // moved to pending, but not executed yet
-      manager[method](
-        { text: "query", tags: [], sources: [] },
-        { [method]: fn },
-        {},
-      );
+      manager[method]({ text: "query", tags: [] }, { [method]: fn }, {});
       expect(manager.ctx.status).toBe(TaskStatus.PENDING);
       await vi.advanceTimersByTimeAsync(debounceMs - 50);
       expect(manager.ctx.status).toBe(TaskStatus.PENDING);
@@ -54,24 +50,12 @@ describe("ExecutionManager", () => {
       expect(fn).toHaveBeenCalledTimes(1);
 
       // multiple calls are ignored and only the last one is executed
-      manager[method](
-        { text: "query", tags: [], sources: [] },
-        { [method]: fn },
-        {},
-      );
+      manager[method]({ text: "query", tags: [] }, { [method]: fn }, {});
       expect(manager.ctx.status).toBe(TaskStatus.PENDING);
       await vi.advanceTimersByTimeAsync(debounceMs - 50);
-      manager[method](
-        { text: "query", tags: [], sources: [] },
-        { [method]: fn },
-        {},
-      );
+      manager[method]({ text: "query", tags: [] }, { [method]: fn }, {});
       await vi.advanceTimersByTimeAsync(debounceMs - 50);
-      manager[method](
-        { text: "query", tags: [], sources: [] },
-        { [method]: fn },
-        {},
-      );
+      manager[method]({ text: "query", tags: [] }, { [method]: fn }, {});
       await vi.advanceTimersByTimeAsync(debounceMs);
       expect(manager.ctx.status).toBe(TaskStatus.COMPLETE);
       expect(fn).toHaveBeenCalledTimes(1 + 1);
@@ -128,18 +112,10 @@ describe("ExecutionManager", () => {
 
       expect(manager.ctx.status).toBe(TaskStatus.INITIAL);
 
-      manager[method](
-        { text: "query", tags: [], sources: [] },
-        { [method]: fn },
-        {},
-      );
+      manager[method]({ text: "query", tags: [] }, { [method]: fn }, {});
       expect(manager.ctx.status).toBe(TaskStatus.PENDING);
       await vi.advanceTimersByTimeAsync(debounceMs + 10);
-      manager[method](
-        { text: "query", tags: [], sources: [] },
-        { [method]: fn },
-        {},
-      );
+      manager[method]({ text: "query", tags: [] }, { [method]: fn }, {});
 
       await vi.advanceTimersByTimeAsync(LATENCY + 1);
 
@@ -162,7 +138,7 @@ describe("ExecutionManager", () => {
     expect(manager.ctx.status).toBe(TaskStatus.INITIAL);
 
     // immediately transition to pending
-    manager.search({ text: "query", tags: [], sources: [] }, operations, {});
+    manager.search({ text: "query", tags: [] }, operations, {});
     expect(manager.ctx.status).toBe(TaskStatus.PENDING);
 
     // finish search, back to complete
@@ -170,15 +146,15 @@ describe("ExecutionManager", () => {
     expect(manager.ctx.status).toBe(TaskStatus.COMPLETE);
 
     // keep staying pending if search frequent enough
-    manager.search({ text: "query", tags: [], sources: [] }, operations, {});
+    manager.search({ text: "query", tags: [] }, operations, {});
     expect(manager.ctx.status).toBe(TaskStatus.PENDING);
     await vi.advanceTimersByTimeAsync(SEARCH_DEBOUNCE_MS - 50);
     expect(manager.ctx.status).toBe(TaskStatus.PENDING);
-    manager.search({ text: "query", tags: [], sources: [] }, operations, {});
+    manager.search({ text: "query", tags: [] }, operations, {});
     expect(manager.ctx.status).toBe(TaskStatus.PENDING);
     await vi.advanceTimersByTimeAsync(SEARCH_DEBOUNCE_MS - 50);
     expect(manager.ctx.status).toBe(TaskStatus.PENDING);
-    manager.search({ text: "query", tags: [], sources: [] }, operations, {});
+    manager.search({ text: "query", tags: [] }, operations, {});
     await vi.advanceTimersByTimeAsync(SEARCH_DEBOUNCE_MS - 50);
     expect(manager.ctx.status).toBe(TaskStatus.PENDING);
 
@@ -187,20 +163,20 @@ describe("ExecutionManager", () => {
     expect(manager.ctx.status).toBe(TaskStatus.COMPLETE);
 
     // debounce timer resets since we do 'ask' after 'search'
-    manager.search({ text: "query", tags: [], sources: [] }, operations, {});
+    manager.search({ text: "query", tags: [] }, operations, {});
     await vi.advanceTimersByTimeAsync(SEARCH_DEBOUNCE_MS - 50);
     expect(manager.ctx.status).toBe(TaskStatus.PENDING);
-    manager.ask({ text: "query", tags: [], sources: [] }, operations, {});
+    manager.ask({ text: "query", tags: [] }, operations, {});
     await vi.advanceTimersByTimeAsync(ASK_DEBOUNCE_MS - 50);
     expect(manager.ctx.status).toBe(TaskStatus.PENDING);
     await vi.advanceTimersByTimeAsync(ASK_DEBOUNCE_MS - 50);
     expect(manager.ctx.status).toBe(TaskStatus.COMPLETE);
 
-    manager.ask({ text: "query", tags: [], sources: [] }, operations, {});
+    manager.ask({ text: "query", tags: [] }, operations, {});
     expect(manager.ctx.status).toBe(TaskStatus.PENDING);
     await vi.advanceTimersByTimeAsync(ASK_DEBOUNCE_MS - 50);
     expect(manager.ctx.status).toBe(TaskStatus.PENDING);
-    manager.ask({ text: "query", tags: [], sources: [] }, operations, {});
+    manager.ask({ text: "query", tags: [] }, operations, {});
     expect(manager.ctx.status).toBe(TaskStatus.PENDING);
     await vi.advanceTimersByTimeAsync(ASK_DEBOUNCE_MS + 50);
     expect(manager.ctx.status).toBe(TaskStatus.COMPLETE);